// src/dto/enums.ts
// This file was generated by a custom prisma generator, do not edit manually.
export const Role = {
	ADMIN: 'ADMIN',
	USER: 'USER',
	HOST: 'HOST',
	MANAGER: 'MANAGER',
} as const;

export type Role = (typeof Role)[keyof typeof Role];

export const OrderStatus = {
	AWAITING: 'AWAITING',
	RECEIVED: 'RECEIVED',
	SERVED: 'SERVED',
	CANCELED: 'CANCELED',
	DISPUTED: 'DISPUTED',
	COMPLETED: 'COMPLETED',
} as const;

export type OrderState = (typeof OrderStatus)[keyof typeof OrderStatus];

export const SpicyLevel = {
	NOT_SPICY: 'NOT_SPICY',
	MILD: 'MILD',
	MEDIUM: 'MEDIUM',
	HOT: 'HOT',
	EXTRA_HOT: 'EXTRA_HOT',
} as const;

export type SpicyLevel = (typeof SpicyLevel)[keyof typeof SpicyLevel];

export const PaymentStatus = {
	NONE: 'NONE',
	PAID: 'PAID',
	REFUNDED: 'REFUNDED',
	CANCELLED: 'CANCELLED',
	PENDING: 'PENDING',
} as const;

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus];

export const DrinkTemperature = {
	COLD: 'COLD',
	ROOM: 'ROOM',
	HOT: 'HOT',
} as const;

export type DrinkTemperature = (typeof DrinkTemperature)[keyof typeof DrinkTemperature];

export const TableStatus = {
	AVAILABLE: 'AVAILABLE',
	OCCUPIED: 'OCCUPIED',
	RESERVED: 'RESERVED',
	ORDERING: 'ORDERING',
	SERVING: 'SERVING',
	PAYMENT: 'PAYMENT',
} as const;

export type TableStatus = (typeof TableStatus)[keyof typeof TableStatus];

export const OrderActionType = {
	CREATE: 'CREATE',
	UPDATE: 'UPDATE',
	ADD_ITEM: 'ADD_ITEM',
	REMOVE_ITEM: 'REMOVE_ITEM',
	CHANGE_STATUS: 'CHANGE_STATUS',
} as const;

export type OrderActionType = (typeof OrderActionType)[keyof typeof OrderActionType];

export const PaymentType = {
	CASH: 'CASH',
	CREDIT_CARD: 'CREDIT_CARD',
	DEBIT_CARD: 'DEBIT_CARD',
} as const;

export type PaymentType = (typeof PaymentType)[keyof typeof PaymentType];

export const Allergies = {
	NONE: 'NONE',
	GLUTEN: 'GLUTEN',
	DAIRY: 'DAIRY',
	EGG: 'EGG',
	PEANUT: 'PEANUT',
	TREENUT: 'TREENUT',
	FISH: 'FISH',
	SHELLFISH: 'SHELLFISH',
	SOY: 'SOY',
	SESAME: 'SESAME',
	CELERY: 'CELERY',
	MUSTARD: 'MUSTARD',
	LUPIN: 'LUPIN',
	SULPHITES: 'SULPHITES',
	MOLLUSCS: 'MOLLUSCS',
} as const;

export type Allergies = (typeof Allergies)[keyof typeof Allergies];

export const FoodType = {
	APPETIZER: 'APPETIZER',
	MAIN_COURSE: 'MAIN_COURSE',
	DESERT: 'DESERT',
	SIDES: 'SIDES',
	SOUSE: 'SOUSE',
	OTHER: 'OTHER',
} as const;

export type FoodType = (typeof FoodType)[keyof typeof FoodType];

export const FoodCategory = {
	SALAD: 'SALAD',
	MEAT: 'MEAT',
	SOUP: 'SOUP',
	FISH: 'FISH',
	VEGGIES: 'VEGGIES',
	OTHER: 'OTHER',
	SEAFOOD: 'SEAFOOD',
} as const;

export type FoodCategory = (typeof FoodCategory)[keyof typeof FoodCategory];

export const DrinkCategory = {
	BEER: 'BEER',
	WINE: 'WINE',
	SPIRITS: 'SPIRITS',
	COFFEE: 'COFFEE',
	TEA: 'TEA',
	OTHER: 'OTHER',
	SODA: 'SODA',
	ALCOHOLIC: 'ALCOHOLIC',
	NON_ALCOHOLIC: 'NON_ALCOHOLIC',
} as const;

export type DrinkCategory = (typeof DrinkCategory)[keyof typeof DrinkCategory];